Fixed rewrite, everything should work again
