+++ src/States.hs
@@ -1,6 +1,7 @@
-{-# LANGUAGE TemplateHaskell, TypeSynonymInstances, FlexibleInstances #-}
+{-# LANGUAGE TemplateHaskell #-}
 module States where
 
+import Brick
 import Brick.Widgets.FileBrowser
 import Brick.Widgets.List (List)
 import Data.Map.Strict (Map)
@@ -9,7 +10,8 @@ import Lens.Micro.Platform
 import System.Random.MWC (GenIO)
 import Stack hiding (head)
 import Types
-import qualified Data.Map.Strict as Map
+import qualified Data.List.NonEmpty as NE
+import qualified Data.Map.Strict as M
 import qualified Stack
 
 type Name = ()
@@ -22,9 +24,6 @@ data Mode  = MainMenu
            | FileBrowser 
            | Cards
   deriving (Show, Eq, Ord)
-
--- class HasMode t where
---   getMode :: t -> Mode
 
 data State = MainMenuState     MMS
            | SettingsState     SS
@@ -70,6 +69,30 @@ data CardState =
   , _number         :: Int
   }
 
+defaultCardState :: Card -> CardState
+defaultCardState Definition{} = DefinitionState { _flipped = False }
+defaultCardState (MultipleChoice _ _ ics) = MultipleChoiceState 
+  { _highlighted = 0
+  , _number = length ics + 1
+  , _tried = M.fromList [(i, False) | i <- [0..length ics]] }
+defaultCardState (OpenQuestion _ perforated) = OpenQuestionState 
+  { _gapInput = M.empty
+  , _highlighted = 0
+  , _number = nGapsInPerforated perforated
+  , _entered = False
+  , _correctGaps = M.fromList [(i, False) | i <- [0..nGapsInPerforated perforated - 1]] }
+defaultCardState (MultipleAnswer _ answers) = MultipleAnswerState 
+  { _highlighted = 0
+  , _selected = M.fromList [(i, False) | i <- [0..NE.length answers-1]]
+  , _entered = False
+  , _number = NE.length answers }
+defaultCardState (Reorder _ elements) = ReorderState
+  { _highlighted = 0
+  , _grabbed = False
+  , _order = M.fromList (zip [0..] (NE.toList elements))
+  , _entered = False
+  , _number = NE.length elements }
+
 data CS = CS
   { _cards          :: [Card]     -- list of flashcards
   , _index          :: Int        -- current card index
@@ -80,31 +103,21 @@ data CS = CS
   , _showControls   :: Bool
   -- , _incorrectCards :: [Int]      -- list of indices of incorrect answers
   }
--- instance HasMode CS where
---   getMode = const Cards
 
 newtype MMS = MMS 
   { _l  :: List Name String }
--- instance HasMode MMS where
---   getMode = const MainMenu
 
 type IS = ()
--- instance HasMode IS where
---   getMode = const Info
 
 type Settings = Map Int Bool
 
 type SS = (Int, Settings)
--- instance HasMode SS where
---   getMode = const Settings
 
 data CSS = CSS
   { _list       :: List Name String
   , _exception  :: Maybe String
   , _recents    :: Stack FilePath
   }
--- instance HasMode CSS where
---   getMode = const CardSelector
 
 data FBS = FBS
   { _fb          :: FileBrowser Name
@@ -113,8 +126,6 @@ data FBS = FBS
   , _filePath    :: Maybe FilePath
   , _showHidden  :: Bool
   }
--- instance HasMode FBS where
---   getMode = const FileBrowser
 
 getMode :: State -> Mode
 getMode (MainMenuState     _) = MainMenu
@@ -136,13 +147,7 @@ getState :: GlobalState -> State
 getState = fromJust . safeGetState
 
 updateState :: GlobalState -> State -> GlobalState
--- updateState gs s@(MainMenuState _)     = gs & states %~ Map.insert MainMenu s
--- updateState gs s@(SettingsState _)     = gs & states %~ Map.insert Settings s
--- updateState gs s@(InfoState _)         = gs & states %~ Map.insert Info s
--- updateState gs s@(CardSelectorState _) = gs & states %~ Map.insert CardSelector s
--- updateState gs s@(FileBrowserState _)  = gs & states %~ Map.insert FileBrowser s
--- updateState gs s@(CardsState _)        = gs & states %~ Map.insert Cards s
-updateState gs s = gs & states %~ Map.insert (getMode s) s
+updateState gs s = gs & states %~ M.insert (getMode s) s
 
 updateMMS :: GlobalState -> MMS -> GlobalState
 updateMMS gs s = updateState gs (MainMenuState s)
@@ -166,19 +171,7 @@ updateFBS :: GlobalState -> FBS -> GlobalState
 updateFBS gs s = updateState gs (FileBrowserState s)
 
 goToState :: GlobalState -> State -> GlobalState
--- goToState gs s@(MainMenuState _)     = gs & states %~ Map.insert MainMenu s
---                                           & stack  %~ insert MainMenu
--- goToState gs s@(SettingsState _)     = gs & states %~ Map.insert Settings s
---                                           & stack  %~ insert Settings
--- goToState gs s@(InfoState _)         = gs & states %~ Map.insert Info s
---                                           & stack  %~ insert Info
--- goToState gs s@(CardSelectorState _) = gs & states %~ Map.insert CardSelector s
---                                           & stack  %~ insert CardSelector
--- goToState gs s@(FileBrowserState _)  = gs & states %~ Map.insert FileBrowser s
---                                           & stack  %~ insert FileBrowser
--- goToState gs s@(CardsState _)        = gs & states %~ Map.insert Cards s
---                                           & stack  %~ insert Cards
-goToState gs s = gs & states %~ Map.insert (getMode s) s
+goToState gs s = gs & states %~ M.insert (getMode s) s
                     & stack  %~ insert (getMode s)
 
 popState :: GlobalState -> GlobalState
@@ -186,10 +179,14 @@ popState gs = let
   s    = gs ^. stack
   top  = Stack.head s
   s'   = Stack.pop s in
-    gs & states %~ Map.delete top
+    gs & states %~ M.delete top
        & stack  .~ s'
 
 safeGetState :: GlobalState -> Maybe State
 safeGetState gs = do
   key <- safeHead (gs ^. stack)
-  Map.lookup key (gs ^. states)
+  M.lookup key (gs ^. states)
+
+goToModeOrQuit :: GlobalState -> Mode -> EventM n (Next GlobalState)
+goToModeOrQuit gs mode = 
+  maybe (halt gs) (continue . goToState gs) $ M.lookup mode (gs ^. states)
