+++ src/UI/CardSelector.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE TemplateHaskell #-}
 module UI.CardSelector 
   ( State
   , drawUI
@@ -6,42 +5,26 @@ module UI.CardSelector
   , theMap
   , getRecents
   , getRecentsFile
-  , addRecent
-  , runCardSelectorUI ) where
+  , addRecent ) where
 
 import Brick
 import Brick.Widgets.Border
 import Brick.Widgets.Border.Style
 import Brick.Widgets.Center
 import Control.Exception (displayException, try)
-import Control.Monad (filterM)
 import Control.Monad.IO.Class
-import Data.Functor (void)
-import Data.List (sort)
 import Lens.Micro.Platform
 import Parser
 import Recents
-import System.Environment (lookupEnv)
-import System.FilePath ((</>), splitFileName, dropExtension, splitPath, joinPath)
+import Runners
 import States
-import Types
 import UI.Attributes hiding (theMap)
 import UI.BrickHelpers
-import UI.FileBrowser (runFileBrowserUI)
-import UI.Cards (Card, runCardsUI, runCardsWithOptions)
 import qualified Brick.Widgets.List as L
 import qualified Data.Vector as Vec
 import qualified Graphics.Vty as V
 import qualified Stack as S
 import qualified UI.Attributes as A
-
-runCardSelectorUI :: GlobalState -> IO GlobalState
-runCardSelectorUI gs = do
-  rs <- getRecents
-  let prettyRecents = shortenFilepaths (S.toList rs)
-  let options = Vec.fromList (prettyRecents ++ ["Select file from system"])
-  let initialState = CSS (L.list () options 1) Nothing rs
-  return $ gs `goToState` CardSelectorState initialState
 
 drawUI :: CSS -> [Widget Name]
 drawUI s = 
@@ -98,7 +81,7 @@ handleEvent gs s@CSS{_list=l, _exception=exc} (VtyEvent ev) =
                         V.EvKey V.KEnter [] ->
                           case L.listSelectedElement l' of
                             Nothing -> continue' s'
-                            Just (_, "Select file from system") -> continue =<< liftIO (runFileBrowser (update s'))
+                            Just (_, "Select file from system") -> continue =<< liftIO (runFileBrowserUI (update s'))
                             Just (i, _) -> do
                                 let fp = (s' ^. recents) `S.unsafeElemAt` i
                                 fileOrExc <- liftIO (try (readFile fp) :: IO (Either IOError String))
@@ -108,18 +91,10 @@ handleEvent gs s@CSS{_list=l, _exception=exc} (VtyEvent ev) =
                                     Left parseError -> continue' (s' & exception ?~ errorBundlePretty parseError)
                                     Right result -> continue =<< liftIO (do
                                       s'' <- addRecentInternal s' fp
-                                      let s''' = s'' & exception .~ Nothing
-                                      runCardsWithOptions (update s''') result)
-                                      -- return $ update ()
+                                      runCardsWithOptions (update s'') result)
                         _ -> continue' s'
 
 handleEvent gs _ _ = continue gs
-
-runFileBrowser :: GlobalState -> IO GlobalState
-runFileBrowser gs = do
-  result <- runFileBrowserUI gs
-  -- maybe (return s) (\(cards, fp) -> addRecentInternal s fp <* runCardsWithOptions (s^.gs) cards) result
-  return result
 
 refreshRecents :: CSS -> IO CSS
 refreshRecents s = do
