+++ src/UI/MainMenu.hs
@@ -1,33 +1,17 @@
-{-# LANGUAGE TemplateHaskell #-}
-module UI.MainMenu (State (..), drawUI, handleEvent, theMap, runMainMenuUI) where
+module UI.MainMenu (State (..), drawUI, handleEvent, theMap) where
 
 import Brick
 import Brick.Widgets.Border
 import Brick.Widgets.Border.Style
 import Brick.Widgets.Center
 import Control.Monad.IO.Class
-import Data.Functor (($>))
 import Lens.Micro.Platform
+import Runners
 import States
 import UI.Attributes
 import UI.BrickHelpers
-import UI.CardSelector (runCardSelectorUI)
-import UI.Info (runInfoUI)
-import UI.Settings (runSettingsUI)
-import qualified Data.Vector as Vec
 import qualified Graphics.Vty as V
 import qualified Brick.Widgets.List as L
-
-runMainMenuUI :: GlobalState -> GlobalState
-runMainMenuUI gs = 
-  let options = Vec.fromList 
-                  [ "Select"
-                  , "Info"
-                  , "Settings"
-                  , "Quit" ]
-
-      initialState = MMS (L.list () options 1) in
-  gs `goToState` MainMenuState initialState
 
 title :: Widget Name
 title = withAttr titleAttr $
